refactor(transformers): Refactor top stock raw data enrichment transformer.
- Rename transformers.py to top_stock_raw_data_enrichment_transformer.py.
- Move top stock raw data enrichment transformer to its own folder.
- Create cloud_interactions.py to handle cloud interactions.
- Update .gitignore.
