+++ workflows/transformers/top_stock_raw_data_enrichment_transformer/config.py
