+++ workflows/transformers/top_stock_raw_data_enrichment_transformer/top_stock_raw_data_enrichment_transformer.py
